// src/api/gen/models/ProgressionPoliciesSearchRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProgressionPolicyFilter } from "./ProgressionPolicyFilter";
export type ProgressionPoliciesSearchRequest = {
  filter?: ProgressionPolicyFilter;
  page?: number;
  size?: number;
};